fix(permissions): fix global permissions by actually implementing them
the code that previously existed didn't really make sense,
but we didn't notice because it was never called.

the solution we provide here is a bit hacky, by encoding what's usually
`app_label.codename` as `auth.app_label_and_codename`, where
`app_label_and_codename` is just app_label and codename concatenated together
with a dot. The reason for this is that we can't go around creating contenttypes
with fixed names as the previous code assumed, that API just doesn't exist in Django.
So we have to refer to one contenttype that encodes "globality" once, and encode
what's usually app_name and codename in the codename together.